A version of Rage is implemented
When the player starts holding the rage button (currently LT), an AoE around the player appears and damages enemies in the area.  The AoE disappears when the rage button is deleted.  I plan to use some timers to turn this into a damage over time area instead of just being able to spam the button for big damage