Dash implementation
When the dash button is first pressed, the player will move rapidly in the currently faced direction for a fixed period of time.  During the dash, collisions with enemies and their projectiles will be ignored.  If the rage button was held at the start of the dash, the player will leave a trail of flames behind them during the dash, dealing AoE damage to enemies in their path.