Archers have normal arrows
Starting when the player walks within a certain range of an archer, the archer will fire arrows at their location when the player is in their line of sight.  The arrows can hit walls or the player, and will deal damage before despawning.  Also a lot of other files got changed as part of adding 2D animated sprite to MC